ID команды теперь беззнаковое целое

+++ KAMEditor/machinetool/components/commandManager/commands/command.h
@@ -15,10 +15,10 @@ public:
     // скорее всего, временные методы, будет переделано при рефакторинге
     virtual std::string getEditorColor() const { return "#000000"; }
     virtual QString getArguments() const = 0;
-    int getId() const { return id; }
+    size_t getId() const { return id; }
 
 private:
-    int id;
+    size_t id;
 };
 
 #endif // COMMAND_H
